Job updated

+++ src/include/os.h
@@ -3,10 +3,10 @@
 
 struct module_os_s {
     int (*filepart)(const char *filename, size_t offset, size_t maxread,
-                    char **dst, size_t *ndst);
+                    char **dst, uint64_t *ndst);
     int (*fileparts)(const char *filename, size_t maxread,
                      unsigned int *parts);
-    int (*filesize)(const char *filename, size_t *size);
+    int (*filesize)(const char *filename, uint64_t *size);
     int (*filewrite)(const char *fname, const char *mode,
                      char *content, int ncontent);
     int (*filejoin)(struct config_s *cfg, const char *fname, char *received,
